+++ Sources/MintCLI/Commands/BootstrapCommand.swift
@@ -6,6 +6,9 @@ import SwiftCLI
 
 class BootstrapCommand: MintfileCommand {
 
+    @Key("-o", "--overwrite", description: "Automatically overwrite a symlinked executable that is not installed by mint without asking. Either (y/n)")
+    var overwrite: Bool?
+
     init(mint: Mint) {
         super.init(mint: mint,
                    name: "bootstrap",
@@ -14,6 +17,6 @@ class BootstrapCommand: MintfileCommand {
 
     override func execute() throws {
         try super.execute()
-        try mint.bootstrap(link: link)
+        try mint.bootstrap(link: link, overwrite: overwrite)
     }
 }